+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-28 10:25:13
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-05-29 15:41:46
+// Commit: style(ui): harmonize light and dark theme palette for ui
 
 class uiComponentState {
-  final String commitRef = '1543';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-05-28T10:25:13+05:00');
+  final String commitRef = '1586';
+  final String lastMessage = 'style(ui): harmonize light and dark theme palette for ui';
+  final DateTime timestamp = DateTime.parse('2026-05-29T15:41:46+05:00');
 }